Trait matrixgraph::BasicGraphMethods

pub trait BasicGraphMethods {
    fn size(&self) -> usize;
    fn get_edge(&self, edge: (usize, usize)) -> Option<f64>;
    fn set_edge(&mut self, edge: (usize, usize), value: Option<f64>);
}

Provides basic graph methods.

Required Methods

fn size(&self) -> usize

Returns the size of the graph, i.e. its number of vertices.

fn get_edge(&self, edge: (usize, usize)) -> Option<f64>

Returns the value of an edge of the graph.

fn set_edge(&mut self, edge: (usize, usize), value: Option<f64>)

Sets the value of an edge of the graph.